Added an ID component to spacemonster subclasses (0 for blueDwarf, 1 for eventHorizon, etc.).
Relied on this rather than "GetType" in the EnemySpawner class.

Changed the UpdateEnemies method into a method that only takes a single SpaceMonster and desubscribes/subscribes to that depending on whether it's being added or removed.

Instead of using "Type", the CallForMateAction class and others just send their SpaceMonster as a parameter, and then check the ID of it to see if it matches the index of the proper list in the enemyDirectoryList.